refactor: remove more code duplication from (dis)connect operations
