+++ src/txmempool.cpp
@@ -478,16 +478,7 @@ void CTxMemPool::addAddressIndex(const CTxMemPoolEntry &entry, const CCoinsViewC
         uint8_t address_type{0};
         uint160 address_bytes;
 
-        if (prevout.scriptPubKey.IsPayToScriptHash()) {
-            address_type  = AddressType::P2SH;
-            address_bytes = uint160(TrimScriptP2SH(prevout.scriptPubKey));
-        } else if (prevout.scriptPubKey.IsPayToPublicKeyHash()) {
-            address_type  = AddressType::P2PKH;
-            address_bytes = uint160(TrimScriptP2PKH(prevout.scriptPubKey));
-        } else if (prevout.scriptPubKey.IsPayToPublicKey()) {
-            address_type  = AddressType::P2PK;
-            address_bytes = Hash160(TrimScriptP2PK(prevout.scriptPubKey));
-        } else {
+        if (!AddressBytesFromScript(prevout.scriptPubKey, address_type, address_bytes)) {
             continue;
         }
 
@@ -503,16 +494,7 @@ void CTxMemPool::addAddressIndex(const CTxMemPoolEntry &entry, const CCoinsViewC
         uint8_t address_type{0};
         uint160 address_bytes;
 
-        if (out.scriptPubKey.IsPayToScriptHash()) {
-            address_type  = AddressType::P2SH;
-            address_bytes = uint160(TrimScriptP2SH(out.scriptPubKey));
-        } else if (out.scriptPubKey.IsPayToPublicKeyHash()) {
-            address_type  = AddressType::P2PKH;
-            address_bytes = uint160(TrimScriptP2PKH(out.scriptPubKey));
-        } else if (out.scriptPubKey.IsPayToPublicKey()) {
-            address_type  = AddressType::P2PK;
-            address_bytes = Hash160(TrimScriptP2PK(out.scriptPubKey));
-        } else {
+        if (!AddressBytesFromScript(out.scriptPubKey, address_type, address_bytes)) {
             continue;
         }
 
@@ -570,18 +552,8 @@ void CTxMemPool::addSpentIndex(const CTxMemPoolEntry &entry, const CCoinsViewCac
         uint8_t address_type{0};
         uint160 address_bytes;
 
-        if (prevout.scriptPubKey.IsPayToScriptHash()) {
-            address_type  = AddressType::P2SH;
-            address_bytes = uint160(TrimScriptP2SH(prevout.scriptPubKey));
-        } else if (prevout.scriptPubKey.IsPayToPublicKeyHash()) {
-            address_type  = AddressType::P2PKH;
-            address_bytes = uint160(TrimScriptP2PKH(prevout.scriptPubKey));
-        } else if (prevout.scriptPubKey.IsPayToPublicKey()) {
-            address_type  = AddressType::P2PK;
-            address_bytes = Hash160(TrimScriptP2PK(prevout.scriptPubKey));
-        } else {
-            address_type  = AddressType::UNKNOWN;
-            address_bytes.SetNull();
+        if (!AddressBytesFromScript(prevout.scriptPubKey, address_type, address_bytes)) {
+            continue;
         }
 
         CSpentIndexKey key = CSpentIndexKey(input.prevout.hash, input.prevout.n);
